+++ src/components/ButtonComponent/ButtonComponent.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Button } from 'reactstrap';
+
+const ButtonComponent: React.FC = () => {
+    return (
+        <div>
+            <Button
+                color="primary">Click Me
+            </Button>
+        </div>
+    );
+};
+
+export default ButtonComponent;
