--- src/components/Table/index.ts
@@ -1,3 +1,0 @@
-import Table from './Table';
-
-export default Table;
