+++ src/components/TableTemplate/index.ts
@@ -1,0 +1,3 @@
+import TableTemplate from './TableTemplate';
+
+export default TableTemplate;
